lol, removing graphQL
