--- graphql/src/main/resources/META-INF/services/io.strimzi.admin.http.server.registration.RouteRegistration
@@ -1,1 +1,0 @@
-io.strimzi.admin.graphql.GraphQLService
