--- graphql/src/main/resources/graphql-schema/baseSchema.graphql
@@ -1,7 +1,0 @@
-schema {
-    query: Query
-    mutation: Mutation
-}
-
-type Mutation {}
-type Query {}
