--- graphql/src/main/resources/log4j2.properties
@@ -1,11 +1,0 @@
-name = GraphQL
-
-appender.console.type = Console
-appender.console.name = STDOUT
-appender.console.layout.type = PatternLayout
-appender.console.layout.pattern = %d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
-
-rootLogger.level = ${env:GRAPHQL_LOG_LEVEL:-INFO}
-rootLogger.appenderRefs = stdout
-rootLogger.appenderRef.console.ref = STDOUT
-rootLogger.additivity = false
