--- kafka-admin/src/main/resources/META-INF/services/io.strimzi.admin.graphql.registration.GraphQLRegistration
@@ -1,1 +1,0 @@
-io.strimzi.admin.kafka.admin.KafkaAdminService
